Update session callback
Cookies are being set on requests regardless of the fact that the session was never accessed. This was causing me problems because I have a cookie path set to something other than what my static resources live under. So all calls to all static resources were generating a new cookie which was being sent back to my client, clobbering the real cookie.